refactor search index
